Forked heredoc + lexer norminette
